requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.11
django-cors-headers==3.11.0
django-debug-toolbar==3.2.1
django-guardian==2.4.0
djangorestframework==3.12.4
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
